+++ src/routes/Friends.js
@@ -1,5 +1,7 @@
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import FriendsScreen from ".";
+import { Text } from "native-base";
+import FriendsScreen from "../screens/Friends";
+
 const NavFriendStack = createNativeStackNavigator();
 
 const FriendsStackScreen = () => {
@@ -8,6 +10,9 @@ const FriendsStackScreen = () => {
       initialRouteName="Friends"
       screenOptions={{
         tabBarShowLabel: false,
+        headerLeft: () => { return <Text fontWeight="bold" fontSize="26px" pl={4}>Friends</Text> },
+        headerShadowVisible: false,
+        title: "",
       }}
     >
       <NavFriendStack.Screen name="Friends" component={FriendsScreen} />
