+++ src/screens/Assistance/AssistanceStackScreen.js
@@ -1,26 +1,26 @@
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import AssistanceScreen from ".";
-import AssistanceForm from "./AssistanceForm";
-import AssistanceList from "./AssistanceList";
+// import { createNativeStackNavigator } from "@react-navigation/native-stack";
+// import AssistanceScreen from ".";
+// import AssistanceForm from "./AssistanceForm";
+// import AssistanceList from "./AssistanceList";
 
-const AssistanceStack = createNativeStackNavigator();
+// const AssistanceStack = createNativeStackNavigator();
 
-const AssistanceStackScreen = () => {
-    return (
-      <AssistanceStack.Navigator
-        initialRouteName="Assistance"
-        screenOptions={{
-          tabBarShowLabel: false,
-          style: {
-            backgroundColor: "#fff",
-          },
-        }}
-      >
-      <AssistanceStack.Screen name="Assistance" component={AssistanceScreen} />
-      <AssistanceStack.Screen name="AssistanceForm" component={AssistanceForm} />
-      <AssistanceStack.Screen name="AssistanceList" component={AssistanceList} />
-      </AssistanceStack.Navigator>
-    );
-  };
+// const AssistanceStackScreen = () => {
+//     return (
+//       <AssistanceStack.Navigator
+//         initialRouteName="Assistance"
+//         screenOptions={{
+//           tabBarShowLabel: false,
+//           style: {
+//             backgroundColor: "#fff",
+//           },
+//         }}
+//       >
+//       <AssistanceStack.Screen name="Assistance" component={AssistanceScreen} />
+//       <AssistanceStack.Screen name="AssistanceForm" component={AssistanceForm} />
+//       <AssistanceStack.Screen name="AssistanceList" component={AssistanceList} />
+//       </AssistanceStack.Navigator>
+//     );
+//   };
 
-export default AssistanceStackScreen;
+// export default AssistanceStackScreen;
