+++ src/screens/Assistance/index.js
@@ -6,28 +6,28 @@ import AssistanceLists from "./AssistanceList";
 import { contentLayout, color } from "../../../Style";
 
 const style = StyleSheet.create({
-    page: {
-      alignSelf:"stretch",
-      alignItems:"center"
-    },
-    content: {
-      width: contentLayout.width,
-      overflow:"scroll",
-    },
-    text: {
-      fontSize:"16px",
-      color:"#8E8E8E"
-    },
-    button: {
-      borderRadius:"20",
-      backgroundColor:"#188FFC",
-      width:"100px",
-      height:"35px",
-    },
-    textBtn: {
-      fontSize:"16px",
-      color:"#FFFFFF"
-    }
+  page: {
+    alignSelf: "stretch",
+    alignItems: "center"
+  },
+  content: {
+    width: contentLayout.width,
+    overflow: "scroll",
+  },
+  text: {
+    fontSize: "16pt",
+    color: "#8E8E8E"
+  },
+  button: {
+    borderRadius: "20",
+    backgroundColor: "#188FFC",
+    width: "100pt",
+    height: '35pt',
+  },
+  textBtn: {
+    fontSize: "16pt",
+    color: "#FFFFFF"
+  }
 });
 
 const AssistanceScreen = ({ navigation }) => {
@@ -35,11 +35,11 @@ const AssistanceScreen = ({ navigation }) => {
   return (
     <View style={style.page}>
       <Button
-				style={style.button}
-				leftIcon={<IconFe name="plus" size="sm" color={color.white} />}
-				onPress={() => navigation.navigate(AssistanceForm)} >
-				<Text style={style.textBtn}>Create</Text>
-			</Button>
+        style={style.button}
+        leftIcon={<IconFe name="plus" size="sm" color={color.white} />}
+        onPress={() => navigation.navigate(AssistanceForm)} >
+        <Text style={style.textBtn}>Create</Text>
+      </Button>
       <Center style={style.content}>
         <IconFe name="x-circle" size="50px" color={color.grey} />
         <Text style={style.text}>You haven't created any appointments.</Text>
